Creates connection.go to hold all papergres db connection context
